requirements/common.txt
-e git://github.com/brutimus/django.git@b79e0fbedd83d4e03cacc95d7b309038adc9c090#egg=Django
dj-database-url==0.3.0
dj-static==0.0.5
django-databrowse==1.3
django-debug-toolbar==1.2
gunicorn==18.0
pystache==0.5.3
pytz==2014.2
sqlparse==0.1.11
static==1.0.2
South==0.8.4
wsgiref==0.1.2